more accurate aiming of the lasers

+++ api/buildcraft/api/power/ILaserTarget.java
@@ -33,9 +33,9 @@ public interface ILaserTarget {
 	 */
 	boolean isInvalidTarget();
 
-	int getXCoord();
+	double getXCoord();
 
-	int getYCoord();
+	double getYCoord();
 
-	int getZCoord();
+	double getZCoord();
 }
